requirements.txt
anyio==3.7.0
bcrypt==4.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
crypto==1.4.1
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
msgpack==1.0.5
Naked==0.1.32
passlib==1.7.4
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==0.21.1
PyYAML==6.0
requests==2.31.0
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
zipp==3.15.0
